refactor(language-hub-mobile): make animation reusable
